refactor(scraper): simplify `Scraper` interface
