+++ packages/scraper/src/amazon.ts
@@ -1,17 +1,19 @@
+import * as cheerio from 'cheerio';
 import { parseCurrency, parseNum, parseURL } from './lib/parse';
 import { findBySelectors } from './lib/utils';
-import ProductScraper from './scraper';
-import { Image, Product, Scraper } from './types';
+import { Product, Scraper } from './types';
 
 export default class AmazonScraper implements Scraper {
-  private scrapper: ProductScraper;
+  $: cheerio.CheerioAPI;
+  url: URL;
 
   constructor(url: string, html: string) {
-    this.scrapper = new ProductScraper(url, html);
+    this.$ = cheerio.load(html);
+    this.url = new URL(url);
   }
 
-  getHTMLData(): Product {
-    const name = findBySelectors(this.scrapper.$, [
+  getProduct(): Product {
+    const name = findBySelectors(this.$, [
       {
         selector: '#productTitle',
         attribute: 'value',
@@ -22,7 +24,7 @@ export default class AmazonScraper implements Scraper {
       },
     ]);
 
-    const price = findBySelectors(this.scrapper.$, [
+    const price = findBySelectors(this.$, [
       {
         selector: '#priceValue',
         attribute: 'value',
@@ -33,7 +35,7 @@ export default class AmazonScraper implements Scraper {
       },
     ]);
 
-    const currency = findBySelectors(this.scrapper.$, [
+    const currency = findBySelectors(this.$, [
       {
         selector: '#currencyOfPreference',
         attribute: 'value',
@@ -44,20 +46,20 @@ export default class AmazonScraper implements Scraper {
       },
     ]);
 
-    const image = findBySelectors(this.scrapper.$, [
+    const image = findBySelectors(this.$, [
       { selector: 'img#landingImage', attribute: 'src' },
       { selector: '#imgTagWrapperId img', attribute: 'src' },
       { selector: '[name="productImageUrl"]', attribute: 'value' },
     ]);
 
-    const category = findBySelectors(this.scrapper.$, [
+    const category = findBySelectors(this.$, [
       {
         selector: '#productCategory',
         attribute: 'value',
       },
     ]);
 
-    const imageURL = parseURL(image, this.scrapper.hostname);
+    const imageURL = parseURL(image, this.url.hostname);
 
     const product = {
       name,
@@ -70,17 +72,5 @@ export default class AmazonScraper implements Scraper {
     };
 
     return JSON.parse(JSON.stringify(product));
-  }
-
-  getMetadata() {
-    return this.scrapper.getMetadata();
-  }
-
-  getJsonLd() {
-    return this.scrapper.getJsonLd();
-  }
-
-  getMicrodata() {
-    return this.scrapper.getMicrodata();
   }
 }
