+++ packages/scraper/src/lib/__tests__/utils.test.ts
@@ -1,23 +1,8 @@
 import { Product } from '@/types';
-import {
-  getHostname,
-  mergeProducts,
-  normalizeText,
-  findBySelectors,
-  getJsonLdProduct,
-  getJsonLdCurrency,
-  getJsonLdPrice,
-} from '../utils';
 import * as cheerio from 'cheerio';
+import { findBySelectors, mergeProducts, normalizeText } from '../utils';
 
 describe('Utils Module', () => {
-  describe('getHostname', () => {
-    it('should return the hostname without www', () => {
-      expect(getHostname('https://www.example.com')).toBe('www.example.com');
-      expect(getHostname('https://example.com')).toBe('example.com');
-    });
-  });
-
   describe('mergeProducts', () => {
     it('should combine product details', () => {
       const products: Product[] = [
@@ -92,45 +77,6 @@ describe('Utils Module', () => {
       const $ = cheerio.load(html);
       const selectors = [{ selector: 'img', attribute: ['data-src', 'src'] }];
       expect(findBySelectors($, selectors)).toEqual('src.jpg');
-    });
-  });
-
-  describe('getJsonLdProduct', () => {
-    it('should return the product from JSON-LD', () => {
-      const jsonLd = { '@graph': [{ '@type': 'Product', name: 'Product 1' }] };
-      expect(getJsonLdProduct(jsonLd)).toEqual({
-        '@type': 'Product',
-        name: 'Product 1',
-      });
-    });
-
-    it('should return undefined if no product is found', () => {
-      const jsonLd = { '@graph': [{ '@type': 'Person', name: 'John Doe' }] };
-      expect(getJsonLdProduct(jsonLd)).toBeUndefined();
-    });
-  });
-
-  describe('getJsonLdCurrency', () => {
-    it('should return the currency from JSON-LD', () => {
-      const data = { offers: { priceCurrency: 'USD' } };
-      expect(getJsonLdCurrency(data)).toBe('USD');
-    });
-
-    it('should return undefined if no currency is found', () => {
-      const data = { offers: {} };
-      expect(getJsonLdCurrency(data)).toBeUndefined();
-    });
-  });
-
-  describe('getJsonLdPrice', () => {
-    it('should return the price from JSON-LD', () => {
-      const data = { offers: { price: 100 } };
-      expect(getJsonLdPrice(data)).toBe(100);
-    });
-
-    it('should return undefined if no price is found', () => {
-      const data = { offers: {} };
-      expect(getJsonLdPrice(data)).toBeUndefined();
     });
   });
 });
